[Comet] Set status bar height to match specs and tweak cutout rect
The height specs of the status bar are 56dp when the camera is at the
top and 36dp when the camera is at the bottom.
To achieve this, we set the status bar height config to 36dp, which will
be used when the camera is at the bottom.
Then to achieve 56dp when the camera is at the top, we use the cutout
rect approximation, and make it 56dp in height. The height of the
status bar is adjusted to be at least as tall as the cutout rect.

+ Tweak the cutout rect, to make it be in the corner and intersect with
  the top and right of the screen, which is needed for status bar inset
  calculations. Before, it had a vertical offset, and was offscreen to
  the right.
  Now it is a rect that starts from the top right of the screen and goes
  to 56dp in height and until the end of the camera in width.

Test: Manual
Fixes: 309773362
Change-Id: I9425364bab2e1413c02674cb2ae2c5be411f70c3