[Comet] Tweak cutout rect approximation to be 56p in width and height
The specs say that the status bar and nav bar should be 56dp tall when
they contain the cutout. For the status bar this is when the cutout is
at the top (rotation 0 and rotation 90). For the nav bar this is when
the cutout is at the bottom (rotation 180 and rotation 270).

The status bar and nav bar have logic to make their height to be at
least as tall as the cutout rect approximation.

Fixes: 309773362
Bug: 299908906
Test: Manual
Change-Id: I1c01280558b145280449a8c199f680c1e4304ee6