[Comet] Tweak outer display cutout approximation rect to fix insets
Since the cutout rect was not defined to start at the top of the screen,
it was not considered to be on the side edge when the device was in
landscape/seascape.
This would make the status bar insets calculation not take it into
account, and content would be displayed over the cutout area.

Test: Manual
Fixes: 321973422
Change-Id: I8a47f4636cf7251ff6c1f9e4e274c94ca44aacaf