Update PMU thresholds with PMU ignorance logic disabled
pa/2830448 removes the core ignorance logic in PMU and causes the
behavior of PMU to change. With the new PMU logic, we need to re-tune
all the SPC thresholds for camera use-cases. This CL adjusts SPC
thresholds accordingly, so that we do not have power regression once the
PMU change in the kernel is merged.

Refer to ag/27702721

Bug: 341495529
Test: Build pass, Boot to home
Test: With pa/2830448 and 30fps(ag/27631025)/60fps(ag/27725456) RT changes, it will increase ~40mW without frame drop impact to improve performace issue.
30fps/120fps test results (https://docs.google.com/spreadsheets/d/1yhYV8LUaraO8mt1cYNFhXj-QzNBkTTtc3xXh-ryH2Eo/edit?usp=sharing&resourcekey=0-xs3Qtufdju8eYOZDTAC7oQ)
60fps/240fps test results (https://docs.google.com/spreadsheets/d/1uiG4jCR9rsLuD7WseiaptpktbUAt5BLRnyN3Teg8JII/edit?usp=sharing&resourcekey=0-lBJxBct6NT_rtkp1brC4Vg)
Change-Id: Id7bd8704039d8a7f35bee93c491327fed3101211