powerhint: Set response_time_ms for clusters
By default slow it down as auto margins can lead to faster ramping up in
low activity use cases.

Set it back to default in GAME mode to retain previous behavior. Set it
explicitly for DISPLAY_IDLE as a potential area for more tuning in the
future.

Bug: 315748906
Flag: EXEMPT powerhints are not compatible with flagging
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I275545cc98714afd3efefb9f57e6bac55fcb6ea3